Scripts: Add test-e2e script
