+++ package.json
@@ -90,13 +90,12 @@
 		"jest-puppeteer": "3.2.1",
 		"jsdom": "11.12.0",
 		"lerna": "3.4.3",
-		"lint-staged": "7.2.0",
+		"lint-staged": "7.3.0",
 		"lodash": "4.17.10",
 		"mkdirp": "0.5.1",
 		"node-sass": "4.11.0",
 		"pegjs": "0.10.0",
 		"phpegjs": "1.0.0-beta7",
-		"puppeteer": "1.6.1",
 		"react-dom": "16.6.3",
 		"react-test-renderer": "16.6.3",
 		"redux": "4.0.0",
@@ -179,7 +178,7 @@
 		"publish:prod": "npm run build:packages && lerna publish",
 		"test": "npm run lint && npm run test-unit",
 		"pretest-e2e": "concurrently \"./bin/reset-e2e-tests.sh\" \"npm run build\"",
-		"test-e2e": "cross-env JEST_PUPPETEER_CONFIG=test/e2e/puppeteer.config.js wp-scripts test-unit-js --config test/e2e/jest.config.json --runInBand",
+		"test-e2e": "wp-scripts test-e2e --config test/e2e/jest.config.json",
 		"test-e2e:watch": "npm run test-e2e -- --watch",
 		"test-php": "npm run lint-php && npm run test-unit-php",
 		"test-unit": "wp-scripts test-unit-js --config test/unit/jest.config.json",
