+++ packages/scripts/scripts/test-e2e.js
@@ -1,0 +1,1 @@
+require( './test-e2e-jest-puppeteer' );
